Fix penalty assignment when offense is set wrong
If a penalty can only be committed by the offense/defense, then that
team is automatically assigned. This was the case even if the team on
offense was unknown (for example due to
Ten-Point-Touchdowns/Football-Data-Converter#12). Now, if the team is
unknown, other methods are tried first to determine the correct team.